pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = "testBank nub"
authors = ["selivanov"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "~0.15.0"
fastapi = "~0.70.0"
alembic = "~1.7.4"
SQLAlchemy = "~1.4.26"
aiosqlite = "~0.17.0"
databases = "~0.5.3"
pytest = "~6.2.5"
requests = "~2.26.0"
email-validator = "~1.1.3"
python-multipart = "~0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
SQLAlchemy-Utils = "^0.37.9"
emails = "^0.6"
Jinja2 = "^3.0.2"
